requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
tensorflow
Pillow
opencv-python
numpy
